+++ ft_calloc.c
@@ -12,7 +12,7 @@
 
 #include "libft.h"
 
-void	*ft_calloc(unsigned long nmemb, unsigned long size)
+void	*ft_calloc(size_t nmemb, size_t size)
 {
 	void	*s;
